Fix looking up icons
What the current code does is looking for the best match, storing it
into QString path...and then overwrite it with empty string on next run
if the icon is not found in the current dir. So we need to check first
if the icon is available in the theme and only then store the path,
otherwise empty string gets returned non-deterministically.

This also fixes case of some missing icons in Plasma Next.

REVIEW: 118415
BUG: 330845