Allow passing custom theme base path together with custom app themes
KIconTheme in kdelibs4 was searching
KGlobal::dirs()->resourceDirs("data") to find all icons; the status notifier dataengine was then setting custom resourceDirs(..) with custom SNI theme paths (which the SNIs can pass) and so the SNI icons always ended up in the theme search paths (the SNI icons are stored as a whole theme).

With the port to QStandardPaths, we lost the ability to pass custom dirs
into the theme search paths and that results in status notifier icons in
Plasma Next having "unknown" icons as their icon theme paths are not
searched and so icons are not found. This is the case mostly of the Qt4
apps running on Qt with the QSystrayIcon-to-SNI-patches (case of *buntu
and I heard opensuse too?)

KIconLoader however has "addAppDir(..)" method, so I expanded that
method to actually take an "app dir" and add it to the theme search
paths. Plasma Next now have proper icons in the systray.

REVIEW: 118561