Optimize KIconEngine::availableSizes()
It's called a ton of times (over 30k times when opening kdevelop and
changing area at some point) and it's quite slow nowadays. The reason is
that we're retrieving the icon path every time to see if it's available
(because it goes through aaaaaaall of the themes' directories to find the
available icon).

This patch adds a dictionary in the KIconLoader so that we can cache for
the icons we already found, clears it when the theme changes.
Also we're construct the same sizes list every time, now we're always
returning the same list.

REVIEW: 123924