set default theme to breeze rather than oxygen
Other framework defaults use Breeze (colors, style..) but icons were still
oxygen, making theming in Qt native applications seem off when using the
frameworkintegartion plugin while not running inside a session that was
coerced into Breeze theming by the kde4breeze theme.
This notably applies to rooted applications that operate on a different
$HOME but might still be made to load the KDE platformtheme plugin.

The breeze icon theme internally uses oxygen as a fallback, so even if
an icon is missing in breeze we will continue to find it with breeze
as default.

CHANGELOG: Change default theme from oxygen to breeze to offer consistency.
FIXED-IN: 5.16
BUG: 336739
REVIEW: 125720