KIconLoader: reduce number of lookups when doing fallbacks
In the first pass, we already checked whether the icon exists in all icon themes.
So in the second pass, there's no point in starting with that again, we can go
ahead and start doing fallbacks right away.

(The comment about the two passes was removed in b84858c, but some of it still
applies.)

Tested with: strace -e file ./kiconengine_unittest testUnknownIconNotCached |& wc -l
Before: 6341
After: 4589

REVIEW: 128514