Support Icon Scale from Icon naming specification 0.13
Since Qt 5.9 there's a ScaledPixmapHook in QIconEngine which is called when device pixel ratio is > 1 and it wants a scaled pixmap.
In contrast to regular pixmap() this also knows the scale factor.

Implement support for Icon Scale properties introduced in Icon naming specification 0.13 [1] to choose an appropriate icon,
if available, rather than blatantly loading a larger icon that may not fit in the context and also could be hard to tell as
its physical size will be smaller than what it was designed for.

[1] https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-0.13.html

BUG: 365941

Differential Revision: https://phabricator.kde.org/D6313