Start drawing emblems in the bottom-right corner
Summary:
While writing a HIG page about emblems (D15897), we noticed an inconsistency in where they're currently placed: All of our icons that have built-in emblems as well as all 3rd-party system tray icons put their emblems in the bottom-right corner. But when `KIconLoader` it used to add emblems to a regular icon, it puts the first emblem in the bottom-left corner.

This arbitrary inconsistency makes it impossible to write a clear rule in the HIG about emblem placement. Changing hundreds of icons and 3rd-party apps to move all their emblems to the bottom-right corner doesn't seem likely. Therefore, this patch adjusts 'KIconLoader` to position the first emblem in the bottom-right corner to match existing icons and 3rd-party apps.

Test Plan:
Symlink emblem:
{F6303725, size=full}

Shared folder emblem:
{F6303727, size=full}

Emblem on a disk in the Places panel:
{F6303728, size=full}

Folders with multiple emblems:
{F6303730, size=full}

Reviewers: #vdg, dfaure

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D15947