Produce output with the request devicePixelRatio
This worked when we were returning content from the cache, but the initial
pixmap we created on first load had the device pixel ratio of the input
instead, which made the scaling/positioning all wrong.

Caught by the previously added unit test on the CI, as that runs with a
clean cache.