[KIconDialog] Revamp UI
This modernizes the UI of KIconDialog that hasn't changed since childhood.
The code is overall tidied up and layout moved into a `.ui` file.

The default size is changed from 4x4 to 6x3 to accommodate more widescreen
monitors and the grid allows for three lines of text since many icon names
are longer than just one line.

Furthermore, the stark separation between "system" and "user" icons is lifted,
and instead the "Browse..." button is always accessible and "Other" is just
another item in the "context" combo box.

Also, a "No icons" placeholder is shown when there are no icons in the
given category or because of the search term input.

Finally, the icons are now loaded on demand as the view scrolls along rather
than loading all thousands of them at once with a slow progress bar.

BUG: 388807
BUG: 401408