Make icon colouring a per-icon property rather than a system
At the moment we have icon users setting a global property into a
singleton rather than keeping it on a per-icon basis. This only works as
long as you are not dealing with different icons at the same time.

Instead, keep the QPalette in the KIconEngine (which is per-icon
instance) and the icon will know itself which palette to use.
This commit doesn't change the previous behaviour because it's widely
used although I imagine we can port many of the use cases into this one
and clean a bunch of code.

See PlasmaDesktopTheme::iconFromTheme in qqc2-desktop-style and
qqc2-breeze-style.

CCBUG: 442533