KIconLoader: add enum values with the icon size in the name to StdSizes
And add a TODO to deprecate the old named ones in KF6; at which point
porting applications that use this enum can be done without #ifdef's.

We're about to add 256 and 512 to the StdSizes enum, and apparently
SizeGigantic and SizeColossal are a bit weird for enum member names, and in
the future we'd have to be more creative to find adjectives.

Size16, Size32 are easier to use, albeit a tiny bit uglier.