KIconTheme: allow to also fallback to Breeze-dark when set through QPA
KIconTheme always fallbacks to Breeze icon theme in order to make sure
KDE apps are not missing icons on other desktops, like GNOME. Problem is
that we always fallback to the light variant of Breeze, which will have
icons that are barely visible when applications are using dark theme.
This change checks fallback icon theme set through QPlatformTheme, like
QGnomePlatform which can fallback to correct Breeze variant depending on
the used system theme.