Add KIconTheme::initTheme() to configure proper icon set as opt-in
By calling

KIconTheme::initTheme()

before the application creation one can hard override
the icon engine and the theme.

The function will honor the user settings from the application config
with fallback to kdeglobals.

If the application is already using the KDE platform theme, the icon set will not
be touched and the platform theme will ensure proper theming.