Drop theme-dependant icon sizes
Generally the icon size used is something that should be chosen based on the size of UI elements.
Icon themes shouldn't influence that, otherwise there might be a mismatch with the application style.

We also inconsistently apply this vs the KIconLoader constants that are not theme-dependant, causing
more potential inconsistencies.

Since the value used by Breeze/Oxygen for Desktop group is different to the default defined here
adjust the default value so that it doesn't result in a behavior change