qml/icondialog: forward accepted/rejected signals
While the underlying KIconDialog has accepted/rejected signals, they are
not forwarded to the qml version. The existing iconNameChanged signal
can do the job in some circumstances, but does not adequately cover all
circumstances. For example, the custom icon in Kicker is set by an
IconDialog, but the property can be independently reset to be cleared.
If the user sets a custom icon, clears it, then tries to set the same
icon again, it will not work as the dialog's iconName property stays the
same. In this case, we actually need to know whether the user pressed
Accept or Cancel, so we can set the configuration property again.

This change adds the signals and connects them to those for the
underlying KIconDialog.