Prefer themes on disk over qrc
Reasons are that we want on Plasma (and elsewhere) all our apps to use
the same icons and not have inconstincies when for example breeze is updated.
It makes it possible as well to override icons by the user distro
in the normal way otherwise this is plain not possible. This also matches
how configuration works, the compiled in default has the lowest priority.