Drop setting up a KColorSchemeManager
Creating a KColorSchemeManager actively takes over managing the application palette.

KIconThemes is the wrong layer to do that. It should only ever follow the current application
palette, not enforce one on its own.

Whether or not to use KColorSchemeManager for the application should be a decision by the application author,
independent from whether KIconTheme::initTheme() is used