Avoid saving state between multiple usages of IconDialog
If you select an icon, close the dialog, open it again, and
select the same icon, there won't be any signal of a new icon
being selected because the dialog saves the state from the
previous usage.
The user cannot avoid this behavior as the iconName is
readonly, so it can only be re-set by the IconDialog itself.

BUG:460451